View evento modificado.
Se seccionan los eventos en el calendario por 3 tipos (platicas,
reclutamiento y actividades), tambien se cambio como se despliega el
dia de hoy.
El view evento ahora despliega los datos en textViews (scrolleables),
ademas, se agregaron los botones de agregar evento y mandar mail.
La funcion mandar mail funciona al 100%, la de agregar al iCal sigue en
desarrollo.

+++ CeldasMedida/ViewControllerCalendario.m
@@ -78,7 +78,7 @@ GlobalCalendar *myCalendar;
         event = [[myCalendar eventDayList] objectAtIndex:i];
         if(event == dia)
             {
-                _eventIndex = i/2;
+                _eventIndex = (int)i/2;
                 _diaString = dayString;
                 [self performSegueWithIdentifier:@"Event" sender:self];
             }
@@ -94,7 +94,7 @@ GlobalCalendar *myCalendar;
     if ([[segue identifier] isEqualToString:@"Event"]) {
         EventoViewController *evc = [segue destinationViewController];
         evc.diaString = _diaString;
-        evc.eventIndex = (int)_eventIndex;
+        evc.eventIndex = _eventIndex;
     }
 }
